backend/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.0
apify-client==1.6.1
httpx==0.27.0
pillow==10.2.0
ollama==0.1.6
pydantic==2.6.1
websockets==12.0